Added plugin initalization into env.php Renamed Plugins\Plugin to Plugins\CobaltPlugin Started working on the plugin loading mechanism Implemented the plugin route loading in init_route_table Moved list of router tables into a variable, no longer hard-coded Removed the differentiation of routes/ and routes/core since it was redundant Started work on globals/plugins.php
