+++ classes/Plugins/Manager.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Plugin Manager
+ * 
+ * This class manages active plugins
+ * 
+ * @author Gardiner Bryant <gardiner@heavyelement.io>
+ * @license https://github.com/heavyelementinc/cobalt-core/license
+ * @copyright 2021 - Heavy Element, Inc.
+ */
