+++ env.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * env.php - The Cobalt Environment Bootstrapper
  * 
@@ -15,11 +16,12 @@
  * 
  * @license cobalt-core/license
  * @author Gardiner Bryant <gardiner@heavyelement.io>
+ * @copyright 2021 - Heavy Element, Inc.
  */
 
 // Let's make sure we're running a suppoted version of PHP (since we use 
 // the not-insane array syntax and the spread "..." syntax)
-if(!version_compare(PHP_VERSION, "7.4", ">=")) die("You must be running PHP version 7.4 or greater");
+if (!version_compare(PHP_VERSION, "7.4", ">=")) die("You must be running PHP version 7.4 or greater");
 
 // ENV_ROOT defines the root of the core files (the dir this file resides in)
 define("__ENV_ROOT__", __DIR__);
@@ -27,12 +29,13 @@ define("__ENV_ROOT__", __DIR__);
 // Establish our app root
 $app_root = "";
 // Go up one directory so we're not in the public space
-if(!empty($_SERVER['DOCUMENT_ROOT'])) $app_root = $_SERVER['DOCUMENT_ROOT'] . "/../";
+if (!empty($_SERVER['DOCUMENT_ROOT'])) $app_root = $_SERVER['DOCUMENT_ROOT'] . "/../";
 // Rely on the Cobalt CLI to mandate the path to our app
-else if(key_exists("cli_app_root",$GLOBALS)) $app_root = $GLOBALS['cli_app_root'];
+else if (key_exists("cli_app_root", $GLOBALS)) $app_root = $GLOBALS['cli_app_root'];
 else die("Cannot establish absolute path to app root"); // Die.
 
 define("__APP_ROOT__", realpath($app_root));
+define("__PLG_ROOT__", __APP_ROOT__ . "/plugins");
 
 // Define a few values that we will use to handle writing output during an exception
 $allowed_to_exit_on_exception = true;
@@ -43,15 +46,18 @@ require_once __DIR__ . "/globals/global_exceptions.php";
 require_once __DIR__ . "/globals/global_functions.php";
 // Import Composer's autoload
 $composer = __DIR__ . "/vendor/autoload.php";
-if(file_exists($composer)) require_once $composer;
+if (file_exists($composer)) require_once $composer;
 else die("Dependencies have not been installed. Run `composer install` in the cobalt-core directory");
 // And then define our own autoload function (specified in global_functions.php)
-spl_autoload_register("cobalt_autoload",true);
+spl_autoload_register("cobalt_autoload", true);
+
+// Load our ACTIVE plugins.
+require_once __ENV_ROOT__ . "/globals/plugins.php";
 
 // Instantiate our settings (`true` for loading settings from cache)
-try{
+try {
     $application = new SettingsManager(true);
-} catch (Exception $e){
+} catch (Exception $e) {
     die($e->getMessage());
 }
 /** @global $app How we set up and process our settings */
@@ -60,7 +66,9 @@ $app = $application;
 /** @global __APP_SETTINGS__ The __APP_SETTINGS__ constant is an array of app 
  *                           settings 
  * */
-define("__APP_SETTINGS__",$application->get_settings());
+define("__APP_SETTINGS__", $application->get_settings());
+
+
 
 /** If we're NOT in a CLI environment, we should import the context processor */
-if(!defined("__CLI_ROOT__")) require_once __ENV_ROOT__ . "/globals/context.php";
+if (!defined("__CLI_ROOT__")) require_once __ENV_ROOT__ . "/globals/context.php";
